requirement.txt
beautifulsoup4
lxml
markdown2
pycrypto
